moved TObjectWrapper into own header
